pyproject.toml
[project]
name = "shopping-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agents>=1.4.0",
    "chainlit>=2.6.2",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
]
